Travis: Fast finish old PR builds
Using Travis CI's API and information about the current build, this
script checks to see if the current PR build is the most recent one. It
does this by comparing the current PR build's build number to other
build numbers of builds for this PR. If it is not the most recent build
for the PR, then the build is failed stopping it from proceeding
further. However, if it is the most recent build number or if it is not
a PR (e.g. a build on a user's fork), then the build proceeds without
issues.