Store build numbers in a `set`.
Instead of sorting them, store them in a `set`. While still effectively
sorts them, but will ensure there are no duplicates. This is handy as we
may need to insert the current build number just in case no other builds
match this one, but we don't want a duplicate. Not to mention, creating
a `set` compared to sorting a `list` appears to be several times faster
based on limited benchmarking.