Make sure the current build number is included.
We have seen in some cases that the current build number may not be
included by Travis CI's API. In this case, `max` will receive an empty
collection of build numbers, to which it will respond with an error.
While we are unclear as to why this occurs (maybe an infrequent API
bug), a reasonable workaround is to ensure that the current build number
is definitely included in the collection of build numbers. This way
`max` will at least return the current build number, which will allow
the build to proceed without issues.