Bump to 4.4.1
As we have fixed a bug in the fast finish CI PR build script, but not
otherwise changed the API, we simply bump the patch number. That said,
all consumers of this script that we know of are downloading it directly
from `master`. So the new packages are not strictly necessary. Still it
is good to indicate the level of change made here. Thus we bump the
patch number to indicate the API has not otherwise changed.