Pin conda-build to 2.1.4
This pins `conda-build` to the last known working version. For some
reason, newer versions of `conda-build` have issues with creating an
environment in some cases and may be related to an upload issue we are
experiencing. Not to mention some other errors that have come up. This
should hopefully workaround these issues in feedstocks.