Skip adding existing packages under other labels
This appears to be the cause of relabeling `dev` and `rc` packages under
the `main` label. Not sure if there were any use cases targeted by this
behavior. However given it is causing more harm than good, it seems
reasonable to rip this out until a better solution can be devised for
the problem it was intended to solve.