Disable `conda`'s auto-update feature
If we do decide to pin `conda` for some reason, we don't want to have it
subverted by some other `conda` `install` or `update` command. Nor do we
want to have pin `conda` in every `install` or `update` command. To
avoid this we simply disable `conda`'s auto-update feature. That way we
get the version we install unless we explicitly request a different
version.