Write out parameter check completely
Seems there is some bug with Python `dict`'s `get` method, which we are
running into here. In particular, it seems to ignore the `default` value
provided to it. AFAICT it seems to be related to the key we are
providing (maybe length related :/). This additional code acts as a
workaround for this bug to ensure we are able to handle the pre-CircleCI
2.0 case.